refactor: extract `Command` from `ecs` directory
