+++ src/scene.h
@@ -1,9 +1,9 @@
 #pragma once
 
 #include "./collections/deque.h"
-#include "./ecs/command.h"
 #include "./ecs/components.h"
 #include "./ecs/entities.h"
+#include "command.h"
 #include "common.h"
 #include "input.h"
 #include "level_segment.h"
